+++ scripts/restart.sh
@@ -2,4 +2,6 @@
 
 ./scripts/kill_nodeos.sh
 ./scripts/start_nodeos.sh
+
+sleep 2
 ./scripts/deploy.sh