+++ scripts/test.sh
@@ -3,7 +3,7 @@
 # setparams
 # 1 day   = 86400
 # 30 days = 2592000
-cleos push action eosio.wps setparams '[{"vote_margin": 15, "deposit_required": "100.0000 EOS", "voting_interval": 2592000, "max_monthly_budget": "50000.0000 EOS", "min_time_voting_end": 86400}]' -p eosio.wps
+cleos push action eosio.wps setparams '[{"vote_margin": 2, "deposit_required": "100.0000 EOS", "voting_interval": 120, "max_monthly_budget": "1000.0000 EOS", "min_time_voting_end": 10}]' -p eosio.wps
 
 # init
 cleos -v push action eosio.wps init '["1970-01-01T00:00:00"]' -p eosio.wps
@@ -13,7 +13,9 @@ cleos -v push action eosio.wps setproposer '["myaccount", [{"key":"region", "val
 
 # propose draft
 cleos -v push action eosio.wps submitdraft '["myaccount", "mywps", "My WPS", "500.0000 EOS", 3, [{"key":"region", "value":"global"}]]' -p myaccount
-cleos -v push action eosio.wps submitdraft '["toaccount", "towps", "To WPS", "100.0000 EOS", 6, [{"key":"category", "value":"other"}]]' -p toaccount
+cleos -v push action eosio.wps submitdraft '["toaccount", "towps", "To WPS", "200.0000 EOS", 6, [{"key":"category", "value":"other"}]]' -p toaccount
+cleos -v push action eosio.wps submitdraft '["toaccount", "short", "Short WPS", "800.0000 EOS", 1, [{"key":"category", "value":"short"}]]' -p toaccount
+cleos -v push action eosio.wps submitdraft '["toaccount", "novote", "No Vote WPS", "500.0000 EOS", 1, [{"key":"category", "value":"novote"}]]' -p toaccount
 
 # # cancel draft
 # cleos -v push action eosio.wps canceldraft '["myaccount", "mywps"]' -p myaccount
@@ -23,31 +25,41 @@ cleos -v push action eosio.wps submitdraft '["toaccount", "towps", "To WPS", "10
 
 # deposit EOS into account
 cleos -v transfer myaccount eosio.wps "100.0000 EOS" ""
-cleos -v transfer toaccount eosio.wps "300.0000 EOS" ""
+cleos -v transfer toaccount eosio.wps "400.0000 EOS" ""
 
-# refund
+# # refund
 # cleos -v push action eosio.wps refund '["myaccount"]' -p myaccount
 
 # activate
 cleos -v push action eosio.wps activate '["myaccount", "mywps", "1970-01-01T00:00:00"]' -p myaccount
 cleos -v push action eosio.wps activate '["toaccount", "towps", "1970-01-01T00:00:00"]' -p toaccount
+cleos -v push action eosio.wps activate '["toaccount", "short", "1970-01-01T00:00:00"]' -p toaccount
+cleos -v push action eosio.wps activate '["toaccount", "novote", "1970-01-01T00:00:00"]' -p toaccount
 
 # fund wps
-cleos -v transfer eosio.ramfee eosio.wps "50000.0000 EOS" ""
+cleos -v transfer eosio.ramfee eosio.wps "2000.0000 EOS" ""
 
-# cancel
+# # cancel
 # cleos -v push action eosio.wps canceldraft '["myaccount", "mywps"]' -p myaccount
 
 # vote
 cleos -v push action eosio.wps vote '["mybp1", "mywps", "yes"]' -p mybp1
-cleos -v push action eosio.wps vote '["mybp2", "mywps", "no"]' -p mybp2
+cleos -v push action eosio.wps vote '["mybp2", "mywps", "yes"]' -p mybp2
 cleos -v push action eosio.wps vote '["mybp3", "mywps", "yes"]' -p mybp3
 cleos -v push action eosio.wps vote '["mybp4", "mywps", "abstain"]' -p mybp4
-cleos -v push action eosio.wps vote '["mybp5", "mywps", "yes"]' -p mybp5
+cleos -v push action eosio.wps vote '["mybp5", "mywps", "no"]' -p mybp5
 
-cleos -v push action eosio.wps vote '["mybp1", "towps", "yes"]' -p mybp1
+# cleos -v push action eosio.wps vote '["mybp1", "towps", "yes"]' -p mybp1
 cleos -v push action eosio.wps vote '["mybp3", "towps", "yes"]' -p mybp3
+cleos -v push action eosio.wps vote '["mybp4", "towps", "yes"]' -p mybp4
 cleos -v push action eosio.wps vote '["mybp5", "towps", "yes"]' -p mybp5
 
-# complete
-cleos push action eosio.wps complete '["2019-12-25T00:00:00"]' -p eosio.wps
+cleos -v push action eosio.wps vote '["mybp3", "short", "yes"]' -p mybp3
+cleos -v push action eosio.wps vote '["mybp4", "short", "yes"]' -p mybp4
+cleos -v push action eosio.wps vote '["mybp5", "short", "yes"]' -p mybp5
+
+cleos -v push action eosio.wps vote '["mybp4", "novote", "no"]' -p mybp4
+cleos -v push action eosio.wps vote '["mybp5", "novote", "no"]' -p mybp5
+
+# # complete
+# cleos push action eosio.wps complete '[]' -p eosio.wps