+++ src/deposits.cpp
@@ -14,7 +14,7 @@ void wps::refund( const eosio::name account )
     token::transfer_action transfer( "eosio.token"_n, { get_self(), "active"_n });
     transfer.send( get_self(), account, remaining_balance, "refund" );
 
-    // TESTIN PURPOSES
+    // TESTING PURPOSES
     add_transfer("refund"_n, get_self(), account, remaining_balance, "refund" );
 
     // substract deposits