+++ src/settings.cpp
@@ -40,7 +40,7 @@ void wps::add_funding( const eosio::asset quantity )
 {
     const eosio::name ram_payer = get_self();
 
-    auto state = _state.get_or_default();
+    auto state = _state.get();
     state.available_funding += quantity;
     _state.set( state, ram_payer );
 }
@@ -49,7 +49,7 @@ void wps::sub_funding( const eosio::asset quantity )
 {
     const eosio::name ram_payer = get_self();
 
-    auto state = _state.get_or_default();
+    auto state = _state.get();
     state.available_funding -= quantity;
     _state.set( state, ram_payer );
 }